mkdir, profile presence check, bin file path repair

+++ customize.sh
@@ -9,13 +9,15 @@ ui_print "- unzip"
 unzip -o $ZIPFILE -x 'META-INF/*' -d $MODPATH >&2
 
 ui_print "- create work dir"
-mkdir $smartdns_path
+mkdir -p $smartdns_path/bin $smartdns_path/run $smartdns_path/config
 cp $MODPATH/bin/smartdns-aarch64 $smartdns_path/bin/smartdns
-cp $MODPATH/config/smartdns.conf $smartdns_path/config/smartdns.conf
+if [ ! -f "$smartdns_path/config/smartdns.conf" ]; then
+    cp $MODPATH/config/smartdns.conf $smartdns_path/config/smartdns.conf
+fi
 
 ui_print "- set file permission"
 set_perm_recursive $MODPATH 0 0 0755 0644
 set_perm_recursive $smartdns_path 0 0 0755 0644
-set_perm $smartdns_path/smartdns 0 0 0755
+set_perm $smartdns_path/bin/smartdns 0 0 0755
 
 ui_print "- installed"
